# openapi.yaml
openapi: 3.0.0
info:
  title: FediPrint
  version: 0.0.1
  description: FediPrint API
servers:
  - url: http://127.0.0.1:3000
    description: Local Dev-Server
paths:
...
  /api/v1/user/{username}/followers:
    get:
      tags:
        - activitypub
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of user
        - in: query
          name: page
          schema:
            type: number
          required: false
          description: Page of followers
      description: Get activitypub details of user
      responses:
        '200':
          description: ActivityPub data
          content:
            'application/activity+json':
              schema:
                type: object
  /api/v1/user/{username}/following:
    get:
      tags:
        - activitypub
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of user
        - in: query
          name: page
          schema:
            type: number
          required: false
          description: Page of users which are followed
      description: Get activitypub details of user
      responses:
        '200':
          description: ActivityPub data
          content:
            'application/activity+json':
              schema:
                type: object
...